defconfig: enable CONFIG_OPLUS_FEATURE_FUSE_FS_SHORTCIRCUIT
Enable shortcircuited read/write for files
which already exists on your kernel (perks of having a oplus kernel).

When CONFIG_OPLUS_FEATURE_FUSE_FS_SHORTCIRCUIT is enabled, all the reads and writes
to the fuse mount point go directly to the native filesystem
rather than through the fuse daemon. All requests that aren't
read/write still go through the userspace code.

This allows for significantly better performance on read and writes
and the difference between fuse and the native lower filesystem is
negligible.

Change-Id: I03974b62c9e7949db74dcd8012371ff891f461ea
Signed-off-by: Sarthak Roy <sarthakroy2002@gmail.com>